requirements.txt
python-dotenv
streamlit
Jinja2
pydantic
tenacity
pytest
openai